Fix for M2

+++ src/display.h
@@ -9,10 +9,10 @@
 #define SRC_DISPLAY_H_
 
 void displayPercentage(uint32_t heliPercentage);
-void displayYaw(uint32_t yaw);
+void displayYaw(int16_t yaw);
 void displayMeanVal(uint16_t meanVal);
 void displayBlank(void);
-void displayMessage(uint16_t meanVal, uint32_t displayMode, uint32_t heliPercentage, int32_t yaw);
+void displayMessage(uint16_t meanVal, uint32_t displayMode, uint32_t heliPercentage, int16_t yaw);
 
 //Display state enum
 typedef enum {ALTIPER, ALTIMEAN, YAWDEG, BLANK} displayMode_e;
